fix: ajustando criação e atualiação de ordem de serviço.

+++ src/main/java/br/com/gestao_oficina_api/domain/model/OrdemServico.java
@@ -3,6 +3,7 @@ package br.com.gestao_oficina_api.domain.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -20,7 +21,7 @@ public class OrdemServico {
 
     private String status;
 
-    private Double valorEstimado;
+    private BigDecimal valorEstimado;
 
     @ManyToOne
     @JoinColumn(name = "veiculo_id")
